+++ libspider/libspider/graphs/pisdf/Vertex.h
@@ -164,7 +164,7 @@ namespace spider {
              * @brief Get the number of input edges connected to the vertex.
              * @return number of input edges.
              */
-            inline std::uint32_t inputEdgeCount() const;
+            inline std::uint64_t inputEdgeCount() const;
 
             /**
              * @brief A const reference on the array of output edges. Useful for iterating on the edges.
@@ -184,7 +184,7 @@ namespace spider {
              * @brief Get the number of output edges connected to the vertex.
              * @return number of output edges.
              */
-            inline std::uint32_t outputEdgeCount() const;
+            inline std::uint64_t outputEdgeCount() const;
 
             /**
              * @brief Get the subtype of the vertex.
@@ -289,7 +289,7 @@ namespace spider {
             return inputEdgeArray_.at(ix);
         }
 
-        std::uint32_t Vertex::inputEdgeCount() const {
+        std::uint64_t Vertex::inputEdgeCount() const {
             return inputEdgeArray_.size();
         }
 
@@ -301,7 +301,7 @@ namespace spider {
             return outputEdgeArray_.at(ix);
         }
 
-        std::uint32_t Vertex::outputEdgeCount() const {
+        std::uint64_t Vertex::outputEdgeCount() const {
             return outputEdgeArray_.size();
         }
